refactor: /util modules moved to /gen_epix/common and /test
