+++ gen_epix/seqdb/env.py
@@ -3,6 +3,8 @@ import copy
 import traceback
 from typing import Any, Callable, Type
 
+from gen_epix.common.config import AppCfg
+from gen_epix.common.env import BaseAppEnv
 from gen_epix.fastapp import App, BaseService
 from gen_epix.fastapp.repository import BaseRepository
 from gen_epix.seqdb.domain import DOMAIN, enum, model
@@ -26,8 +28,6 @@ from gen_epix.seqdb.services import (
     SystemService,
     UserManager,
 )
-from util.cfg import AppCfg
-from util.env import BaseAppEnv
 
 
 class AppEnv(BaseAppEnv):
