+++ run.py
@@ -1,5 +1,5 @@
 import gzip
-import importlib.resources
+import importlib
 import os
 import pickle
 from datetime import datetime
@@ -11,113 +11,13 @@ import pandas as pd
 import pytest
 import uvicorn
 
+from gen_epix.common.config import AppCfg, ConfigDiscovery
+from gen_epix.common.domain.enum import AppConfigType, AppType, AppTypeSet
+from gen_epix.common.util import generate_ulid
 from gen_epix.fastapp import Domain
 from gen_epix.fastapp.enum import CrudOperation
 from gen_epix.fastapp.repositories.dict.repository import DictRepository
 from gen_epix.fastapp.repositories.sa.repository import SARepository
-from util.cfg import AppCfg
-from util.util import generate_ulid
-
-
-class AppType(Enum):
-    CASEDB = "casedb"
-    SEQDB = "seqdb"
-    OMOPDB = "omopdb"
-    ALL = "all"
-
-
-class AppTypeSet(Enum):
-    ALL = frozenset({AppType.CASEDB, AppType.SEQDB, AppType.OMOPDB})
-
-
-class AppConfigType(Enum):
-    IDPS = "idps"
-    MOCK_IDPS = "mock_idps"
-    NO_AUTH = "no_auth"
-    DEBUG = "debug"
-    NO_SSL = "no_ssl"
-
-
-class ConfigDiscovery:
-    """
-    Config discovery class.
-    Highest priority is the config path in the environment variable.
-    Second is the local config path.
-    Third is the config path in the package.
-    """
-
-    @staticmethod
-    def get_config_path(
-        app_type: str,
-        env_var_substring: str = "",
-        extension: str = "",
-        verbose: bool = False,
-    ) -> str:
-        """
-        Config path picked in the following order:
-        1. Environment variable
-        2. Local config path
-        3. Package config path
-        4. Raise error if not found
-        """
-        path = ConfigDiscovery.get_config_path_from_env(
-            app_type, env_var_substring=env_var_substring, extension=extension
-        )
-        if path:
-            if verbose:
-                print(f"Config path found in environment variable: {path}")
-            return path
-        path = ConfigDiscovery.get_config_path_from_local(app_type, extension=extension)
-        if path:
-            if verbose:
-                print(f"Config path found in local file: {path}")
-            return path
-        path = ConfigDiscovery.get_config_path_from_package(
-            app_type, extension=extension
-        )
-        if path:
-            if verbose:
-                print(f"Config path found in package: {path}")
-            return path
-        raise ValueError(
-            f"Config path not found for app type {app_type}. Please set the environment variable {app_type.upper()}_CONFIG_PATH."
-        )
-
-    @staticmethod
-    def get_config_path_from_env(
-        app_type: str, env_var_substring: str, extension: str = ""
-    ) -> str | None:
-        """Get config path from environment variable, if not return None."""
-        env_var_name = f"{app_type.upper()}_{env_var_substring}"
-        if env_var_name in os.environ:
-            env_config_path = Path(os.environ[env_var_name])
-            if extension:
-                return str(env_config_path / extension)
-            return str(env_config_path)
-        return None
-
-    @staticmethod
-    def get_config_path_from_local(app_type: str, extension: str = "") -> str | None:
-        """Get config path from local file, if not return None."""
-        local_config_path = Path(f"./config/{app_type}")
-        if local_config_path.exists():
-            if extension:
-                return str(local_config_path / extension)
-            return str(local_config_path)
-        return None
-
-    @staticmethod
-    def get_config_path_from_package(app_type: str, extension: str = "") -> str | None:
-        """Get config path from package, if not return None."""
-        with importlib.resources.as_file(
-            importlib.resources.files("gen_epix")
-        ) as package_path:
-            package_config_path = package_path / app_type / "config"
-        if package_config_path.exists():
-            if extension:
-                return str(package_config_path / extension)
-            return str(package_config_path)
-        return None
 
 
 class Run:
@@ -849,7 +749,7 @@ class Run:
         )
 
     def other_general_run_linters(self) -> None:
-        from util.linter import Linter
+        from test.linter import Linter
 
         file_basename = Path(__file__).parent / "test" / "data" / "output" / "linter"
 
@@ -857,7 +757,7 @@ class Run:
         linter.run_all(file_basename=file_basename)
 
     def other_general_run_pylint(self) -> None:
-        from util.linter import Linter
+        from test.linter import Linter
 
         filter_on_codes = {
             "W0102",
@@ -885,7 +785,7 @@ class Run:
             print(line)
 
     def other_general_run_mypy(self) -> None:
-        from util.linter import Linter
+        from test.linter import Linter
 
         filter_on_codes = {
             "no-untyped-def",
@@ -936,31 +836,6 @@ class Run:
         log_parser.to_excel(out_log_excel_file)
         user_journey = log_parser.create_user_journey()
         user_journey.to_pickle(out_user_journey_file)
-
-    def other_seqdb_parse_ncbi_taxonomy(self) -> None:
-        dir = Path.cwd() / ".ete"
-        os.environ["HOME"] = str(dir)
-        os.environ["XDG_DATA_HOME"] = str(dir / "data")
-        os.environ["XDG_CONFIG_HOME"] = str(dir / "config")
-        os.environ["XDG_CACHE_HOME"] = str(dir / "cache")
-        from util.ncbi_taxonomy import parse_ncbi_taxonomy
-
-        parse_ncbi_taxonomy()
-
-    def other_seqdb_parse_alleles(self) -> None:
-        dir = Path.cwd() / ".ete"
-        os.environ["HOME"] = str(dir)
-        from util.wgmlst import parse_alleles
-
-        parse_alleles()
-
-    def other_seqdb_parse_allele_profiles(self) -> None:
-        from util.wgmlst import parse_allele_profiles
-
-        parse_allele_profiles()
-
-    def other_seq_from_distance_matrix(self) -> None:
-        from util.seqs_from_distance_matrix import main
 
         main()
