--- util/version.py
@@ -1,18 +1,0 @@
-"""Retrieve the project version from the pyproject.toml file."""
-
-import tomllib
-
-
-def get_project_version():
-    """Retrieve the project version from the pyproject.toml file.
-    Must be run from the project root directory.
-
-    Returns:
-        str: The version of the project as specified in pyproject.toml.
-    """
-    pyproject_path = "pyproject.toml"
-
-    with open(pyproject_path, "rb") as f:
-        pyproject_data = tomllib.load(f)
-
-    return pyproject_data["project"]["version"]
